requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
dj-config-url
pillow
django-filter
drf-yasg
celery[redis]
django-allauth
stripe